--- entidad_municipal_app/management/__init__.py
@@ -1,1 +1,0 @@
-
