--- entidad_municipal_app/management/commands/__init__.py
@@ -1,1 +1,0 @@
-
